Remote notification helper methods
We aren’t guaranteed that we will be the only service listening for FCM messages. If another service is listening for remote messages our service will never be called. This provides helper methods + cleanup to allow the developer to easily check if remote messages can be handled by Rover and to tell Rover to handle the remote message.