Update to late-model AGP, Android SDK, kotlin, and dependencies. (#478)
Also add github workflows for CI and deploy, and a simple UI to the Example app for opening experiences interactively.

This officially marks the end of the use of gitflow on this repository. The develop branch (and a patch sitting on it that we're not going to deploy) has been retired.